# perf: less aggresive dyn import polling